+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Do not ask confirmation when explicitly using `castor init` command
 * Allow to pass a callback when using the `run_php` function (similar to the `run` function)
 * Add a new `--castor-file` option to specify a custom root file when using the `castor` command
+* Add a `slug()` function to slugify strings
 
 ### Fixes